Money foundation: integer minor units end-to-end (ALW-299 phase A)
Replaces the float numeric->number \`money\` placeholder with a branded
\`moneyMinor\` integer customType + a pure \`core/money\` value helper
(add/sub/sum/allocate/round, basis-point rates/discounts), per ADR-006.
Migrates Catalog price/cost across the full cascade: contract Zod now
takes integer minor units; the UI boundary converts major<->minor via the
existing @workspace/ui/lib/money; seeds/tests fixtures updated.

- core/money.ts is runtime-pure (no db import); carries its own ISO
  exponent map since ui cannot import core (parallel to ui/lib/money).
- migration 0001 recreates products numeric->integer; hand-edited to
  defer_foreign_keys per ADR-007 (fresh-DB-only template, no value
  transform). Migration-safety guard stays green.
- AGENTS.md + .claude/CLAUDE.md money-convention line updated.

The document hub + de-verticalization land in subsequent phases.

+++ packages/contract/src/catalog/products.ts
@@ -7,8 +7,8 @@ const selectProductSchema = z.object({
   sku: z.string(),
   name: z.string(),
   description: z.string().nullable(),
-  price: z.number().nullable(),
-  cost: z.number().nullable(),
+  price: z.number().int().nullable(),
+  cost: z.number().int().nullable(),
   minStockLevel: z.number().nullable(),
   imageUrl: z.string().nullable(),
   createdAt: z.string(),
@@ -26,7 +26,7 @@ export const productsListSchema = z.array(productSchema);
 export const createProductSchema = z.object({
   name: z.string().min(2, "Name must be at least 2 characters"),
   sku: z.string().min(2, "SKU must be at least 2 characters"),
-  price: z.coerce.number().min(0, "Price must be positive").optional(),
+  price: z.coerce.number().int().min(0, "Price must be positive").optional(),
   description: z.string().nullable().optional(),
   imageUrl: z.string().nullable().optional(),
   minStockLevel: z.coerce
@@ -39,7 +39,7 @@ export const createProductSchema = z.object({
 export const updateProductSchema = z.object({
   name: z.string().min(2).optional(),
   sku: z.string().min(2).optional(),
-  price: z.coerce.number().min(0).optional(),
+  price: z.coerce.number().int().min(0).optional(),
   description: z.string().nullable().optional(),
   imageUrl: z.string().nullable().optional(),
   minStockLevel: z.coerce.number().min(0).optional(),
